Split main.go to many files
